AI/Python/requirements.txt
numpy
opencv-contrib-python
imutils
uvicorn
fastapi
azure-storage-blob
python-decouple